Chapter04/virtualenvs/myproject_env/project/django-myproject/requirements/base.txt
Pillow~=5.2.0
mysqlclient~=1.3.0
Django~=2.1.0
django-crispy-forms~=1.7.0
xhtml2pdf~=0.2.3
django-haystack~=2.8.0
whoosh~=2.7.0
sorl-thumbnail~=12.4.0
